# openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: CAKEService API
    description: The greeting service definition.
    version: 0.0.1
paths:
    /api/campaign/v1/create:
        post:
            tags:
                - CAKEService
            operationId: CAKEService_CreateCampaign
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateCampaignRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CreateCampaignResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/user/v1/login:
        post:
            tags:
                - CAKEService
            operationId: CAKEService_Login
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/LoginRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/LoginResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/user/v1/register:
        post:
            tags:
                - CAKEService
            description: Sends a greeting
            operationId: CAKEService_Register
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/RegisterNewAccountRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/RegisterNewAccountResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/voucher/v1/create:
        post:
            tags:
                - CAKEService
            description: Internal api
            operationId: CAKEService_CreateVoucher
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateVoucherRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CreateVoucherResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
components:
    schemas:
        CreateCampaignRequest:
            type: object
            properties:
                name:
                    type: string
                startDay:
                    type: string
                endDay:
                    type: string
                totalSlot:
                    type: string
                vouchers:
                    type: array
                    items:
                        $ref: '#/components/schemas/CreateCampaignRequest_Data'
            description: Create Campaign
        CreateCampaignRequest_Data:
            type: object
            properties:
                voucherId:
                    type: integer
                    format: int32
        CreateCampaignResponse:
            type: object
            properties:
                status:
                    type: integer
                    format: int32
                message:
                    type: string
                data:
                    $ref: '#/components/schemas/CreateCampaignResponse_Data'
                msgKey:
                    type: string
        CreateCampaignResponse_Data:
            type: object
            properties:
                id:
                    type: string
        CreateVoucherRequest:
            type: object
            properties:
                name:
                    type: string
                type:
                    type: string
                value:
                    type: number
                    format: float
            description: Internal api to create voucher / update voucher / get list voucher
        CreateVoucherResponse:
            type: object
            properties:
                status:
                    type: string
                message:
                    type: string
                msgKey:
                    type: string
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
        LoginRequest:
            type: object
            properties:
                userName:
                    type: string
                password:
                    type: string
            description: Login
        LoginResponse:
            type: object
            properties:
                status:
                    type: integer
                    format: int32
                message:
                    type: string
                data:
                    $ref: '#/components/schemas/LoginResponse_Data'
                msgKey:
                    type: string
        LoginResponse_Data:
            type: object
            properties:
                token:
                    type: string
        RegisterNewAccountRequest:
            type: object
            properties:
                fullName:
                    type: string
                account:
                    type: string
                email:
                    type: string
                phoneNumber:
                    type: string
                userName:
                    type: string
                password:
                    type: string
                birthday:
                    type: string
            description: |-
                Publish api
                 Register
        RegisterNewAccountResponse:
            type: object
            properties:
                status:
                    type: integer
                    format: int32
                message:
                    type: string
                data:
                    $ref: '#/components/schemas/RegisterNewAccountResponse_Data'
                msgKey:
                    type: string
        RegisterNewAccountResponse_Data:
            type: object
            properties:
                token:
                    type: string
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
tags:
    - name: CAKEService